+++ test/protocol/http/header/accept.rb
@@ -26,6 +26,10 @@ describe Protocol::HTTP::Header::Accept::MediaRange do
 end
 
 describe Protocol::HTTP::Header::Accept do
+	it "classifies parse errors as invalid headers" do
+		expect(subject::ParseError.new).to be_a(Protocol::HTTP::InvalidHeaderError)
+	end
+	
 	let(:header) {subject.parse(description)}
 	let(:media_ranges) {header.preferred_media_ranges}